detector/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
requests
numpy
tflite-runtime
Pillow